requirements.txt
openpyxl==3.1.2
python-docx==1.1.2
pillow==10.3.0
reportlab-4.2.0